requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
requests==2.28.1
urllib3==1.26.10

motor~=3.0.0
pip~=22.0.4
attrs~=21.4.0
py~=1.11.0
six~=1.16.0
setuptools~=58.1.0
python-dotenv~=0.20.0